AG-13024 Move `initMock` into new freezableMock.ts file
